mach-venturi: manual registration of input keys
-Instead of using ifdef, we have create two sets of
the structs neccessary to register the input keys with
the platform (USA and INT).

-Created an interface, by which we can register the keys
appropriate to the variant (USA or INT).

-We will be using the /efs/buyer_code.dat to determine
what variant we are running on. We assume that if the
buyer_code.dat contains "XAA" then we are USA otherwise
we are INT.

-This is a neccessary first step in unifying the
USA and INT kernels for venturi.