param: implement new reboot mode.
-In venturi, REBOOT_MODE_NONE is used for reboots and
power offs. We need to know if we are rebooting or just powering
on at start up so we know wether or not we need to disable
charging mode. We don't want to get trapped in charging mode.

-So we are using REBOOT_MODE_REBOOTING (which is ignored by
the bootloader) to signify that we are rebooting and
REBOOT_MODE_NONE for power offs.